moving route processor into its own package
